perf(cli): extract help text to lazy cli/help.ts + --version fast path
Move ~200 lines of help-only symbols (visibleLen, wrapHelpLine, buildHelp,
SUBCOMMAND_HELP, showHelp, hasSubcommandHelp, showSubcommandHelp) from
src/cli/parse.ts into a new leaf module src/cli/help.ts. Update index.ts
to dynamic-import help.ts on the --help path and add a single-token
--version/-v fast path that bypasses parseCli entirely.
